# libs/api-client-lib/model/genericErrorModel.model.ts
/**
 * RealWorld Conduit API
 * Conduit API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GenericErrorModelErrors } from './genericErrorModelErrors.model';


export interface GenericErrorModel { 
    errors: GenericErrorModelErrors;
}